refactor: move the user controller to a separate directory
